requirement.txt
django<2.2.5
psycopg2<2.8.3
django-markdownify<0.8.0
django-social-share<1.4.0
django_social_auth<0.7.28
django-taggit<1.2.0
django-crispy-forms<1.8.1
Pillow<6.2.1
python<3.7.0